LayoutTests/media/modern-media-controls/macos-inline-media-controls/macos-inline-media-controls-volume-styles-expected.txt

Testing the MacOSInlineMediaControls computed styles.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS volumeSliderContainerStyle.position is "absolute"
PASS volumeSliderContainerStyle.bottom is "50px"
PASS volumeSliderContainerStyle.width is "81px"
PASS volumeSliderContainerStyle.height is "31px"
PASS volumeSliderContainerStyle.transform is "matrix(0.00000000000000006123233995736766, -1, 1, 0.00000000000000006123233995736766, -25, -25)"
PASS volumeSliderStyle.left is "11px"
PASS volumeSliderStyle.top is "10px"

PASS successfullyParsed is true

TEST COMPLETE
